Reorganise logger
