+++ packages/apollo-links/src/types.ts
@@ -7,17 +7,3 @@ export type Agreement = {
   url: string;
   token: string;
 }
-
-export interface LogFn {
-  /* tslint:disable:no-unnecessary-generics */
-  <T extends object>(obj: T, msg?: string, ...args: any[]): void;
-  (obj: unknown, msg?: string, ...args: any[]): void;
-  (msg: string, ...args: any[]): void;
-}
-
-export type Logger = {
-  error: LogFn;
-  warn: LogFn;
-  info: LogFn;
-  debug: LogFn;
-}
